Tune default test Normal Gamma priors
The Ng prior used to to cause some wacky standard deviations that would
cause some wacky data in the test states, which would cause the
posterior updates to break because you'd get columns of data with one
value, or columns of data with highly variable data that would cause
numerical instability.

I've just tightened the prior so that's not a problem anymore.