START: Oracle.pit -> Oracle.feature_error
Probability integral transform (PIT) does not work for discrete
variables, so I've made a more general trait, SampleError, to capture
the basic idea that we want an error measure for each column compared to
the data that returns an error in [0, 1] and a centroid showing us where
along the CDF the bulk of the error is.

Continuous variables will still use PIT, while discrete variables will
use something else. I haven't gotten that far yet.