Check for odd types in categorical value map
Before the categorical value map would skip inserting anything that was
not and Entry::Other or Entry::Int. It was assumed that the only other
type of Entry that would make it through would me Entry::EmptyCell. I
want to be able to catch other types if a bug is introduced, so I panic
when a non-categorical type is encountered.