Use uniform proposal for Labeler posterior update
The prior was really strong toward informants being both helpful and
knowledgeable, so using that prior to do posterior update on informants
whose data indicated ignorance or deception lead to poor parameters
since the bulk of the posterior wasn't covered by the prior. I noticed
this when all informants, regardless of how bad there were, were
predicted to label correctly. Changing to uniform made the answers line
up almost exactly with the results of the Bernoulli type.

This is a good lesson in why you might not want to use prior MH. I
should look into the other places where this is done and look for
alternatives.