Auto process string categorical CSV
If you generate a codebook with a value map where the values are
strings, and the csv contains those values, when you create a new
engine, the data in the csv will be automatically converted to integers
based in the indices in the value map. If there is no value map, string
values that cannot be converted to ints will cause a parse error.