Otters and dolphins aren't that dissimilar
There was a doc tests where the rowsim WRT swims was compared between an
otter and a beaver, and a dolphin and a beaver. I expected it to alway
be higher between the beaver and otter, but that was only the case most
of the time. For now I just removed the assertion in that example.