Fix bug in new entropy sobol & exact computations
I was usin state.feature(ix).to_mixture() to get the probability
distribution of a feature, but the the method used to convert a ColModel
into a MixtureType uses uniform weights. If the component weights of
that feature are not uniform (they never are), the answer will be wrong.
I ought to plug the hole in ColModel.to_mixture