BUGFIX oracle-mixture logp discrepancy
The problem was that the `single_view_weights` function in interface
utils used the asgn.log_weights function to get the weights, and the
feature_as_mixture function gets its weights from the view itself. The
view, I think is the more proper way to get the weights, but we need to
make sure that the weights transition always happens -- at least on the
last iteration because if the weights are never updated (using gibbs,
or split-merge) we are going to have problems.