BUGFIX gibbs col/row alg updates weights after run
The gibbs alg does not use weights so it does not update the weights.
However, the weights are used by the oracle to compute probabilities. If
there are no updated weights the probabilities could be wrong, or the
probability function could panic from trying to access a weight that is
not there. This update makes the gibbs alg set the weights to the
empirical weights from the assignment count.