Use ColMetadataList for Engine::insert_data
Before, we used a partial_codebook which was a Codebook where every
field other than col_metadata was ignored. It was a pain in the butt to
maintain that whole thing and not use it, so now we just include the
part that we do need.

Also did some clean up.