Fix continuous predict bug
The fmin_bounded method does badly if there are strong means that are
very spread out. The temporary cure is to first do a brute force search
and then to do fmin_bounded inside an interval discovered by the bute
force grid search. This is not great. In the future I'd like to move
from fmin_bounded to something that uses the derivative, like newton's
method or similar. I've added a couple other optimization methods
(Gradient Descent and Newton's Method) for future work.