Use crates rv; use MC not QMC for entropy ests
Sobol QMC was failing bad for long-tailed joint entropies in the
satellites data set. QMC works by forming a box around the 0.99999
probability region of the joint distribution and laying samples down in
it. This is OK for well-behaved distributions, but when we have one low
variance component in the middle of a very high variance component, the
bounding box can get so wide that the low-variance component hardly gets
any samples in it. MC doesn't have this problem at the expense of it
being slower to compute and to converge -- and it being a random
estimate.