Compute state loglike at end of each MCMC step
Also found a bug where the default categorical hyper can cause `Inf`s in
the MH propsal likelihood. If categorical cardinally (k) is high, the
gamma shape is high; if the rate is one, we have a high probability of
being closer and closer the singularity at 0. If we get close to the
singularity the likelihood is Inf. This propagates to the logprior
diagnostic, which means we can't compute the score. Also, this is
probably really bad for inference, because as soon as the sampler gets
an Inf, it will never move.