predict uncertainty now performs as expected
JSD was used, but it was JS between the final predictive distribution
(averaged over states) and each for the mixture components of that
distribution. Now it is the JSD between the averaged predictive
distribution and the state predictive distributions.

Also note that we've pulled in an upadte to rv which fixes entropy
computation in continuous mixture models.