mnar.missing -> mnar.present; add --no-column-assign
Changed the name of the MissingNotAtRandom `missing` field to `present`
because the value of a cell is `true` if the value is present.

Also added the `-R/--no-column-assign` flag to the braid run CLI command,
which is useful if you want to intitailize a flat column structure and
keep it flat by not running the column reassignment kernel.