Remove metadata encryption and user profiles
Users can encrypt their own metadata. No need for us to reinvent the
wheel. And since the main use case of profiles was to store encryption
keys, we don't need UserInfo. This also allow us to get rid of a good
amount of other junk related to saving and loading things.