fix bug converting dataframes to lace-ready objs
Converting polars and pandas dataframes was done with .to_numpy then
.to_list. .to_numpy is a bit tricky because it wall convert a dataframe
of ints and floats into a array of floats. It there is an object
(string) column in the dataframe, it will convert to an array of object,
which is fine. We never ran into this issue because animals is all int,
which gets converted to an int array, and satellites is floats and
string, which gets converted into an object array.