Be consistent w/ index column name
Previously: lace required the index column to be names a case-
insensitive variant of "ID", but pylace required "index". Now there
is a function in lace_utils that checks whether a string is "ID" or
"index" so either is allowed. This also reduces repetition of `== "id"`
and such, which is error prone.