Regenerated the `index.hbs` theme file to handle mobile sites
Something changed with the MDBook theme that caused it to not render properly on mobile.
Specifically, the site would come up with the sidebar open and the "close" button would not work.
I fixed this by initializing a NEW MDBook, copying over the generated `index.hbs`, and then
  adding in the changes we have made to our `index.hbs` in the Lace Git History.
Now it works correctly on the Firefox Mobile WebDev tool